requirements.txt
pandas==1.1.5
google-cloud==0.34.0
pandas-gbq==0.18.1
gcloud==0.18.3